Make PresentationAdapter generic so it can be reused
